Separate auth cmds
