+++ cmd/login.go
@@ -17,18 +17,13 @@ limitations under the License.
 package cmd
 
 import (
-	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"github.com/dgrijalva/jwt-go"
 	"github.com/fatih/color"
-	"github.com/spf13/viper"
-	"strings"
-
 	cv "github.com/nirasan/go-oauth-pkce-code-verifier"
 	"github.com/skratchdot/open-golang/open"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 	"io"
 	"io/ioutil"
 	"log"
@@ -36,57 +31,11 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"time"
+	"strings"
 )
 
-const clientID = "2VC9z0ZxtzTcQLDNygeEELV3lYFRZwpb"
 const callbackURL = "http://localhost:21900/oauth/callback"
-const domain = "auth.meroxa.io"
 const audience = "https://api.meroxa.io/v1"
-
-func LoginCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "login",
-		Short: "login or sign up to the Meroxa platform",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			err := login()
-			if err != nil {
-				return err
-			}
-			return nil
-		},
-	}
-}
-
-// logoutCmd represents the logout command
-func LogoutCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "logout",
-		Short: "logout of the Meroxa platform",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			// TODO: add confirmation
-			cfg.Set("ACCESS_TOKEN", "")
-			cfg.Set("REFRESH_TOKEN", "")
-			err := cfg.WriteConfig()
-			if err != nil {
-				return err
-			}
-			fmt.Println("credentials cleared")
-			return nil
-		},
-	}
-}
-
-func init() {
-	// Login
-
-}
-
-func login() error {
-	log.Println(color.CyanString("You will now be taken to your browser for authentication or open the url below in a browser."))
-	authorizeUser(clientID, domain, callbackURL)
-	return nil
-}
 
 // AuthorizeUser implements the PKCE OAuth2 flow.
 func authorizeUser(clientID string, authDomain string, redirectURL string) {
@@ -191,6 +140,19 @@ func authorizeUser(clientID string, authDomain string, redirectURL string) {
 	server.Serve(l)
 }
 
+func login() error {
+	log.Println(color.CyanString("You will now be taken to your browser for authentication or open the url below in a browser."))
+	authorizeUser(clientID, domain, callbackURL)
+	return nil
+}
+
+// cleanup closes the HTTP server
+func cleanup(server *http.Server) {
+	// we run this as a goroutine so that this function falls through and
+	// the socket to the browser gets flushed/closed before the server goes away
+	go server.Close()
+}
+
 // getAccessToken trades the authorization code retrieved from the first OAuth2 leg for an access token
 func getAccessTokenAuth(clientID string, codeVerifier string, authorizationCode string, callbackURL string) (string, string, error) {
 	// set the url and form-encoded data for the POST to the access token endpoint
@@ -230,91 +192,16 @@ func getAccessTokenAuth(clientID string, codeVerifier string, authorizationCode 
 	return accessToken, refreshToken, nil
 }
 
-// cleanup closes the HTTP server
-func cleanup(server *http.Server) {
-	// we run this as a goroutine so that this function falls through and
-	// the socket to the browser gets flushed/closed before the server goes away
-	go server.Close()
-}
-
-func refresh(domain string, clientID string, refreshToken string) (accessToken string, newRefreshToken string, err error) {
-	url := fmt.Sprintf("https://%s/oauth/token", domain)
-	var tokenBody = make(map[string]string)
-	tokenBody["client_id"] = clientID
-
-	if refreshToken != "" {
-		tokenBody["grant_type"] = "refresh_token"
-		tokenBody["refresh_token"] = refreshToken
-	} else {
-		return "", "", errors.New("no refresh token")
+func LoginCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "login",
+		Short: "login or sign up to the Meroxa platform",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			err := login()
+			if err != nil {
+				return err
+			}
+			return nil
+		},
 	}
-	requestBody, err := json.Marshal(tokenBody)
-	if err != nil {
-		fmt.Println("marshal:", err)
-		return "", "", err
-	}
-
-	resp, err := http.Post(url, "application/json", bytes.NewBuffer(requestBody))
-	if err != nil {
-		fmt.Println("request error:", err)
-		return "", "", err
-	}
-	defer resp.Body.Close()
-	responseBody, err := ioutil.ReadAll(resp.Body)
-
-	if len(responseBody) > 0 {
-		var ff interface{}
-		json.Unmarshal(responseBody, &ff)
-		result := ff.(map[string]interface{})
-		accessToken = fmt.Sprintf("%v", result["access_token"])
-		if refreshToken == "" {
-			newRefreshToken = fmt.Sprintf("%v", result["refresh_token"])
-		}
-
-		return accessToken, newRefreshToken, nil
-	}
-
-	return "", "", nil
-}
-
-func getAccessToken() (string, error) {
-	// check access token expiration
-	accessToken := cfg.GetString("ACCESS_TOKEN")
-	if accessToken == "" {
-		return "", fmt.Errorf("please login or signup by running 'meroxa login'")
-	}
-
-	// check access exp and grab refresh
-	token, _, err := new(jwt.Parser).ParseUnverified(accessToken, jwt.MapClaims{})
-	if err != nil {
-		fmt.Println(err)
-		return "", fmt.Errorf("please login or signup by running 'meroxa login'")
-	}
-
-	// check token exp
-	claims, ok := token.Claims.(jwt.MapClaims)
-	if !ok {
-		return "", fmt.Errorf("please login or signup by running 'meroxa login'")
-	}
-
-	var exp time.Time
-	tokenExp := claims["exp"].(float64)
-	exp = time.Unix(int64(tokenExp), 0)
-
-	if exp.After(time.Now()) {
-		return accessToken, nil
-	}
-
-	// access token is expire, use refresh
-	refreshToken := cfg.GetString("REFRESH_TOKEN")
-	if refreshToken == "" {
-		return "", fmt.Errorf("please login or signup by running 'meroxa login'")
-	}
-	accessToken, _, err = refresh(domain, clientID, refreshToken)
-	if err != nil {
-		return "", fmt.Errorf("please login or signup by running 'meroxa login'")
-	}
-	cfg.Set("ACCESS_TOKEN", accessToken)
-
-	return accessToken, nil
 }
