+++ cmd/logout.go
@@ -1,0 +1,41 @@
+/*
+Copyright © 2020 Meroxa Inc
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package cmd
+
+import (
+	"fmt"
+	"github.com/spf13/cobra"
+)
+
+// logoutCmd represents the logout command
+func LogoutCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "logout",
+		Short: "logout of the Meroxa platform",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			// TODO: add confirmation
+			cfg.Set("ACCESS_TOKEN", "")
+			cfg.Set("REFRESH_TOKEN", "")
+			err := cfg.WriteConfig()
+			if err != nil {
+				return err
+			}
+			fmt.Println("credentials cleared")
+			return nil
+		},
+	}
+}
